NuGet Cache: ensure for clitool nupkg is there even after skipped 2nd extraction

+++ src/Paket.Core/Dependencies/NuGetCache.fs
@@ -520,7 +520,10 @@ let rec ExtractPackageToUserFolder(source: PackageSource, downloadedNupkgPath:st
                 let newNupkgPath = Path.Combine(parentOfParent.FullName, Path.GetFileName downloadedNupkgPath)
                 File.Move(downloadedNupkgPath, newNupkgPath)
                 do! extract newNupkgPath
-                File.Delete(newNupkgPath)
+                if File.Exists(downloadedNupkgPath) then
+                    File.Delete(newNupkgPath)
+                else
+                    File.Move(newNupkgPath, downloadedNupkgPath)
             else
                 do! extract downloadedNupkgPath
